requirements.txt
numpy-mkl
scipy
sounddevice